@@ -13,16 +13,16 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "9.1.13",
-    "@angular/cdk": "8.2.3",
-    "@angular/common": "9.1.13",
-    "@angular/compiler": "9.1.13",
-    "@angular/core": "9.1.13",
-    "@angular/forms": "9.1.13",
-    "@angular/localize": "^9.1.13",
-    "@angular/platform-browser": "9.1.13",
-    "@angular/platform-browser-dynamic": "9.1.13",
-    "@angular/router": "9.1.13",
+    "@angular/animations": "11.2.14",
+    "@angular/cdk": "11.2.13",
+    "@angular/common": "11.2.14",
+    "@angular/compiler": "11.2.14",
+    "@angular/core": "11.2.14",
+    "@angular/forms": "11.2.14",
+    "@angular/localize": "^11.2.14",
+    "@angular/platform-browser": "11.2.14",
+    "@angular/platform-browser-dynamic": "11.2.14",
+    "@angular/router": "11.2.14",
     "@ngx-translate/core": "^11.0.1",
     "@ngx-translate/http-loader": "^4.0.0",
     "@openecoe/potion-client": "^4.0.12",
@@ -30,34 +30,34 @@
     "autolinker": "^1.8.3",
     "core-js": "^2.6.5",
     "moment": "^2.24.0",
-    "ng-zorro-antd": "9.3.0",
+    "ng-zorro-antd": "11.4.2",
     "ngx-papaparse": "5.1.0",
     "reflect-metadata": "^0.1.13",
     "rxjs": "~6.5.3",
     "socket.io-client": "^2.3.1",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.0",
     "yarn": "^1.22.19",
     "zone.js": "~0.10.2"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.15",
-    "@angular/cli": "9.1.15",
-    "@angular/compiler-cli": "9.1.13",
-    "@angular/language-service": "9.1.13",
-    "@types/jasmine": "^2.8.12",
+    "@angular-devkit/build-angular": "~0.1102.19",
+    "@angular/cli": "11.2.19",
+    "@angular/compiler-cli": "11.2.14",
+    "@angular/language-service": "11.2.14",
+    "@types/jasmine": "~3.6.0",
     "@types/jasminewd2": "^2.0.6",
     "@types/node": "^17.0.29",
-    "codelyzer": "^5.1.2",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
+    "codelyzer": "^6.0.0",
+    "jasmine-core": "~3.8.0",
+    "jasmine-spec-reporter": "~5.0.0",
+    "karma": "~6.4.3",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage-istanbul-reporter": "~3.0.2",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "protractor": "~7.0.0",
     "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "3.8.3"
+    "tslint": "~6.1.0",
+    "typescript": "4.0.8"
   }
 }
