@@ -4,7 +4,7 @@ import {ActivatedRoute, Router} from '@angular/router';
 import {TranslateService} from '@ngx-translate/core';
 import {AuthenticationService} from '../../../services/authentication/authentication.service';
 import {ChronoService} from '../../../services/chrono/chrono.service';
-import { NzModalService } from 'ng-zorro-antd';
+import { NzModalService } from 'ng-zorro-antd/modal';
 
 @Component({
   selector: 'app-state',
@@ -139,7 +139,4 @@ export class StateComponent implements OnInit {
       this.ecoe = value;
     });
   }
-
-
-
 }
