@@ -1,5 +1,4 @@
 import {Pipe, PipeTransform} from '@angular/core';
-import {padStart} from 'ng-zorro-antd';
 
 @Pipe({
   name: 'secondsToMinutes'
@@ -8,6 +7,6 @@ export class SecondsToMinutesPipe implements PipeTransform {
 
   transform(value: number): string {
     const minutes: number = Math.floor(value / 60);
-    return padStart(minutes.toString(), 2, '0') + ':' + padStart((value - minutes * 60).toString(), 2, '0');
+    return minutes.toString().padStart(2, '0') + ':' + (value - minutes * 60).toString().padStart(2, '0');
   }
 }
