@@ -171,17 +171,17 @@ export class QuestionsService {
       }
       const qblock = new Block({name: name, station: station, order: order});
       return qblock.save()
-
+      
     } catch (error) {
       this.logPromisesERROR.push({value: name, reason: error});
       return error;
-
+      
     }
 
+    
 
 
-
-
+    
   }
 
   private async getArea(area: Area | String, ecoe: ECOE | number): Promise<Area> {
@@ -189,39 +189,6 @@ export class QuestionsService {
   }
 
   private async addOptions(questionItem: RowQuestion, idQuestion: number) {
-    // const _schema = new QuestionSchema(questionItem.questionType);
-    //
-    // if (_schema instanceof QuestionBase) {
-    //   _schema.reference = questionItem.reference;
-    //   _schema.description = item[this.HEADER.reference];
-    // }
-    // if (_schema instanceof QuestionRange) {
-    //   _schema.range = item[this.OPTIONS].length;
-    //   _schema.max_points = item[this.HEADER.points];
-    // } else if (_schema instanceof QuestionRadio || _schema instanceof QuestionCheckBox) {
-    //   // _schema.max_points = item[this.HEADER.points];
-    //   const _options = item[this.OPTIONS];
-    //
-    //   for (const { index, value } of _options.map((opt, idx) => ({ idx, opt }))) {
-    //     const _questionOption = new QuestionOption();
-    //
-    //     _questionOption.id_option = index;
-    //     _questionOption.points = value.points;
-    //     _questionOption.label = value.label;
-    //     _questionOption.order = value.order ? value.order : index;
-    //
-    //     _schema.options.push(_questionOption);
-    //   }
-
-
-      // const body = {
-      //   label: ((item[this.OPTION + (idx + 1)])) ? (item[this.OPTION + (idx + 1)]).toString() : item['label'],
-      //   order: (item[this.HEADER.order]) ? item[this.HEADER.order] : idx,
-      //   points: (item[this.POINTS + (idx + 1)]) ? (item[this.POINTS + (idx + 1)]) : item[this.POINTS],
-      //   question: idQuestion
-      // };
-
-
     const savePromises = [];
     const options = questionItem[this.OPTIONS];
     if (options && options.length === 0) {
@@ -279,7 +246,7 @@ export class QuestionsService {
         _schema.description = item[this.HEADER.description];
       }
       if (_schema instanceof QuestionRange) {
-
+        
         const getRateCount = () => {
           const _options = item[this.OPTIONS];
           if (_options) {
@@ -288,13 +255,13 @@ export class QuestionsService {
             return 10;
           }
         }
-
+          
         _schema.range = item[this.HEADER.range] || getRateCount();
-
+        
         _schema.max_points = item[this.HEADER.points];
       } else if (_schema instanceof QuestionRadio || _schema instanceof QuestionCheckBox) {
         const _options = item[this.OPTIONS];
-
+  
         if (_options.length === 0) {
           _options.push(
             new Option({
@@ -303,16 +270,16 @@ export class QuestionsService {
             })
           );
         }
-
+  
         // tslint:disable-next-line:no-shadowed-variable
         for (const { idx, opt } of _options.map((opt, idx) => ({ idx: idx + 1, opt }))) {
           const _questionOption = new QuestionOption();
-
+  
           _questionOption.id_option = idx;
           _questionOption.points = opt.points || opt[`points${idx}`];
           _questionOption.label = opt.label || opt[`option${idx}`];
           _questionOption.order = opt.order ? opt.order : idx;
-
+  
           _schema.options.push(_questionOption);
         }
       }
@@ -325,7 +292,7 @@ export class QuestionsService {
       throw error;
     }
 
-
+    
   }
 
   /**
@@ -337,7 +304,7 @@ export class QuestionsService {
     const _question = new Question();
 
     _question.area = (await this.getArea(rowQuestion[this.HEADER.ac], station.ecoe));
-
+    
     //Si no hay area devolvemos un error y marcamos que no se puede realizar la importación
     if (!_question.area) {
       this.logPromisesERROR.push({
